Move organization files to `components`.
