+++ app/api/v1/api.py
@@ -4,10 +4,10 @@ from app.api.v1.endpoints import geocoding
 from app.api.v1.endpoints import guest_user
 from app.api.v1.endpoints import locations
 from app.api.v1.endpoints import oauth
-from app.api.v1.endpoints import organizations
 from app.api.v1.endpoints import sessions
 from app.api.v1.endpoints import zones
 from app.components.auth import routes as auth
+from app.components.organizations import routes as organizations
 from app.components.user import routes as user
 
 api_router = APIRouter()
