+++ app/components/organizations/models.py
@@ -1,6 +1,5 @@
-from sqlalchemy import Column, String, Integer, DateTime, ForeignKey
+from sqlalchemy import Column, String, Integer, DateTime
 from sqlalchemy.orm import relationship
-from sqlalchemy.sql import func
 
 from app.db.base_class import Base
 from app.db.utc_convertation import utcnow
