+++ app/components/user/crud.py
@@ -3,12 +3,12 @@ from typing import Optional
 
 from sqlalchemy.orm import Session
 
+from app.components.organizations.models import Organization
 from app.components.user.models import User
 from app.components.user.schemas import UserCreate, UserBase, UserInvite
 from app.core.config import settings
 from app.core.security import get_password_hash, verify_password, create_access_token
 from app.crud.crud_oauth import roles
-from app.models.organization import Organization
 
 
 def get_by_email(db: Session, *, email: str) -> Optional[User]:
