+++ app/db/base.py
@@ -1,15 +1,15 @@
-from app.db.base_class import Base
-from app.models.location import Location
-from app.models.reports import Report
-from app.models.reportoptions import ReportOption
-from app.models.changelog import ChangeLog
+from app.components.organizations.models import Organization
 from app.components.user.models import User
-from app.models.sessionhistory import SessionHistory
-from app.models.organization import Organization
+from app.db.base_class import Base
+from app.models.changelog import ChangeLog
 from app.models.geospatial_index import GeospatialIndex
-from app.models.zone import Zone
 from app.models.guest_user import GuestUser
+from app.models.location import Location
 from app.models.oauth import OauthScope, OauthRole, association_table
+from app.models.reportoptions import ReportOption
+from app.models.reports import Report
+from app.models.sessionhistory import SessionHistory
+from app.models.zone import Zone
 
 __all__ = [
     "Base",
