+++ app/db/init_db.py
@@ -1,11 +1,11 @@
 from sqlalchemy.orm import Session
 
 from app import schemas
-
+from app.components import organizations
 # TODO: find a better name
 from app.components import user as userc
+from app.components.organizations import crud as org_crud
 from app.core.config import settings
-from app.crud import crud_organizations as org_crud
 from app.crud.crud_oauth import scopes, roles
 from app.models import oauth
 
@@ -57,7 +57,10 @@ def init_db(db: Session) -> userc.models.User:
     # creating the "DIM" organization
     organization = org_crud.get_by_name(db, "DIM")
     if not organization:
-        organization = org_crud.create(db, obj_in=schemas.OrganizationBase(name="DIM"))
+        organization = org_crud.create(
+            db,
+            obj_in=organizations.schemas.OrganizationBase(name="DIM"),
+        )
 
     # creating first superuser
     user = userc.crud.get_by_email(db, email=settings.FIRST_SUPERUSER)
