+++ app/models/__init__.py
@@ -1,7 +1,6 @@
-from .location import Location
-from .sessionhistory import SessionHistory
-from .reports import Report
-from .reportoptions import ReportOption
 from .changelog import ChangeLog
-from .organization import Organization
 from .geospatial_index import GeospatialIndex
+from .location import Location
+from .reportoptions import ReportOption
+from .reports import Report
+from .sessionhistory import SessionHistory
