+++ app/schemas/__init__.py
@@ -18,8 +18,6 @@ from .location import (
     LocationRequest,
 )
 from .oauth import *
-from .organization import OrganizationBase, OrganizationOut, OrganizationUserInvite
-from .organization import OrganizationBase, OrganizationOut, OrganizationUserInvite
 from .roles import UserRole
 from .session import UserSession
 from .session import UserSession
